Organized Auto Files and Commands into respective folders
